one to one

+++ SpringBootProjects/OneToOne/src/main/java/com/bidirectional/repository/AddressRepository.java
@@ -1,7 +1,7 @@
 package com.bidirectional.repository;
 
-import com.unidirectional.entity.Address;
+import com.bidirectional.entity.Address;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AddressRepository extends JpaRepository<Address, Long> {
+public interface AddressRepository extends JpaRepository<Address, Integer> {
 }
